Split Python and Poetry.
I require Poetry in a specific version, which is not possible when
setting the Python version to be the same as I require (3.11), since it
was built using Python 3.10.
Making use of the latest version and setting my required Python version
works, but requires me to compile Python every time I do a system upgrade
which is super annoying.

Therefore, I decided to treat the two independently and stitch them
together using direnv. This should remove the need of compilation and
improves the caching capabilities of direnv.